requirements.txt
Django==3.2.4
djangorestframework==3.12.4
python-decouple==3.4
django-cors-headers==3.7.0
django-seed==0.2.2